change return type to int for damage

+++ src/main/java/_OOP_develop_gradle/model/Professor.java
@@ -74,7 +74,7 @@ public class Professor {
 	/*
 	 * Ti restituisce danno che il prf causa 
 	 */
-	public double getDamageProf() {
+	public int getDamageProf() {
 		return damage;
 	}
